CLN: move remaning extensions to _libs
