+++ doc/source/whatsnew/v0.20.0.txt
@@ -1230,19 +1230,19 @@ If indicated, a deprecation warning will be issued if you reference theses modul
     "pandas.algos", "pandas._libs.algos", ""
     "pandas.hashtable", "pandas._libs.hashtable", ""
     "pandas.indexes", "pandas.core.indexes", ""
-    "pandas.json", "pandas.io.json.libjson", "X"
-    "pandas.parser", "pandas.io.libparsers", "X"
+    "pandas.json", "pandas._libs.json", "X"
+    "pandas.parser", "pandas._libs.parsers", "X"
     "pandas.formats", "pandas.io.formats", ""
     "pandas.sparse", "pandas.core.sparse", ""
     "pandas.tools", "pandas.core.reshape", ""
     "pandas.types", "pandas.core.dtypes", ""
-    "pandas.io.sas.saslib", "pandas.io.sas.libsas", ""
+    "pandas.io.sas.saslib", "pandas.io.sas._sas", ""
     "pandas._join", "pandas._libs.join", ""
     "pandas._hash", "pandas._libs.hashing", ""
     "pandas._period", "pandas._libs.period", ""
-    "pandas._sparse", "pandas.core.sparse.libsparse", ""
-    "pandas._testing", "pandas.util._testing", ""
-    "pandas._window", "pandas.core.libwindow", ""
+    "pandas._sparse", "pandas._libs.sparse", ""
+    "pandas._testing", "pandas._libs.testing", ""
+    "pandas._window", "pandas._libs.window", ""
 
 
 Some new subpackages are created with public functionality that is not directly
