+++ pandas/_libs/window.pyx
