+++ pandas/util/testing.py
@@ -1476,8 +1476,8 @@ def assert_sp_array_equal(left, right, check_dtype=True):
                              check_dtype=check_dtype)
 
     # SparseIndex comparison
-    assert isinstance(left.sp_index, pd.core.sparse.libsparse.SparseIndex)
-    assert isinstance(right.sp_index, pd.core.sparse.libsparse.SparseIndex)
+    assert isinstance(left.sp_index, pd._libs.sparse.SparseIndex)
+    assert isinstance(right.sp_index, pd._libs.sparse.SparseIndex)
 
     if not left.sp_index.equals(right.sp_index):
         raise_assert_detail('SparseArray.index', 'index are not equal',
